+++ bridges.h
@@ -13,7 +13,7 @@
 #include "uf.h"
 using std::pair, std::vector, std::unordered_map, std::min, std::string;
 vector<pair<long long, long long>> find_bridges(vector<pair<long long, long long>>& edges, size_t num_vertices);
-vector<pair<int, int>> find_bridges_tarjan(vector<pair<int, int>>& edges);
+vector<pair<long long, long long>> find_bridges_tarjan(vector<pair<long long, long long>>& edges);
 double find_bridges_parallel(string& filename, ygm::comm& world);
 void find_bridges_parallel_opt(string& filename, ygm::comm& world);
 void test_disjoint_set(string& csv_file, ygm::comm& world);
